serialisation and process manager tests in a good state
